pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentes-agno"
version = "1.0.0"
dependencies = ["agno", "openai", "python-dotenv", "fastapi", "django", "uvicorn[standard]", "mcp"]

[tool.setuptools.packages.find]
where = ["."]
include = ["agentes_agno*"]
